+++ source/redux/client/render.js
@@ -15,15 +15,16 @@ export default function render({ store })
 	return matchRoutes(store).then(({ redirect, renderArgs }) =>
 	{
 		// In case of a `react-router` `<Redirect/>`
-		if (redirect)
-		{
+		if (redirect) {
 			window.location = getLocationUrl(redirect)
 			throw new Error(`[react-website] (Not an error) Redirecting to ${getLocationUrl(redirect)}`)
-
-			// This kind of a redirect won't work because
-			// the `<Router/>` hasn't been rendered yet.
-			// return store.dispatch(redirect(to))
 		}
+
+		// The first pass of initial client-side render
+		// was to render the markup which matches server-side one.
+		// The second pass will be to render after resolving `getData`.
+		window._react_website_initial_prerender = false
+		window._react_website_skip_preload = false
 
 		// `routes` are used when comparing `instantBack` chain items
 		// for resetting `instantBack` chain when the same route is encountered twice.